Se ordeno el micro de usuarios

+++ usuarios/src/main/java/backend/grupo130/usuarios/data/repository/PostgresUsuarioRepositoryI.java
@@ -4,7 +4,7 @@ package backend.grupo130.usuarios.data.repository;
 import backend.grupo130.usuarios.data.models.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostgresUsuarioRepositoryI extends JpaRepository<Usuario, Integer> {
+public interface PostgresUsuarioRepositoryI extends JpaRepository<Usuario, Long> {
