+++ usuarios/src/main/java/backend/grupo130/usuarios/dto/request/GetByIdRequest.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class GetByIdRequest {
 
-    private final Integer usuarioId;
+    private final Long usuarioId;
 
 }
